[#102]refactor: 필요없는 부분을 삭제하고 변경된 파일 위치를 업데이트
