--- src/pages/Home/types/allObjectItem.ts
@@ -1,6 +1,0 @@
-// child 객체안에 넣을 수 있는 모든 타입을 묶는 타입
-import { ForItem } from "./forItem";
-import { ConditionItem } from "./conditionItem";
-import { PrintItem } from "./printItem";
-import { CodeFlowVariableItem } from "./codeFlow/codeFlowVariableItem";
-export type AllObjectItem = PrintItem | ForItem | ConditionItem | CodeFlowVariableItem;
