--- src/pages/Home/types/dataStructureVarItem.ts
@@ -1,5 +1,0 @@
-// 시각화하는 스택에 들어가는 Var를 정의하는 타입
-import { CodeItem } from "./codeItem";
-export interface DataStructureVarItem extends CodeItem {
-  isLight?: boolean;
-}
