+++ src/pages/Home/types/dataStructuresItem/allDataStructureItem.ts
@@ -1,0 +1,3 @@
+import { DataStructureListItem } from "./dataStructureListItem";
+import { dataStructureVarItem } from "./dataStructureVarItem";
+export type AllDataStructureItem = DataStructureListItem | dataStructureVarItem;
