--- src/pages/Home/types/forItem.ts
@@ -1,17 +1,0 @@
-import { AllObjectItem } from "./allObjectItem";
-export interface ForItem {
-  id: number;
-  start: string;
-  isStartLight: boolean;
-  end: string;
-  isEndLight: boolean;
-  cur: string;
-  isCurLight: boolean;
-  target: string;
-  step: string;
-  isStepLight: boolean;
-  type: string;
-  depth: number;
-  isLight: boolean;
-  child: AllObjectItem[];
-}
