--- src/pages/Home/types/listItem.ts
@@ -1,7 +1,0 @@
-// json 리스트에 들어있는 타입정의
-export interface ListItem {
-  expr: string;
-  highlights: number[];
-  name: string;
-  type: string;
-}
