--- src/pages/Home/types/printItem.ts
@@ -1,12 +1,0 @@
-import { AllObjectItem } from "./allObjectItem";
-
-export interface PrintItem {
-  id: number;
-  type: string;
-  depth: number;
-  isLight: boolean;
-  expr: string;
-  console: string;
-  highlights: number[];
-  child: AllObjectItem[];
-}
