--- src/pages/Home/types/whileItem.ts
@@ -1,11 +1,0 @@
-import { AllObjectItem } from "./allObjectItem";
-
-export interface WhileItem {
-  id: number;
-  type: string;
-  expr: string;
-  highlights: number[];
-  depth: number;
-  isLight: boolean;
-  child: AllObjectItem[];
-}
